move preTask and postTask into extension
