+++ src/WinTenDev.ZiziBot.AppHost/Handlers/NewUpdateHandler.cs
@@ -1,10 +1,7 @@
-﻿using System.Collections.Generic;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using SerilogTimings;
 using Telegram.Bot.Framework.Abstractions;
-using WinTenDev.Zizi.Services.Internals;
 using WinTenDev.Zizi.Services.Telegram;
 using WinTenDev.Zizi.Services.Telegram.Extensions;
 using WinTenDev.Zizi.Utils;
@@ -18,12 +15,7 @@ public class NewUpdateHandler : IUpdateHandler
 
     public NewUpdateHandler(
         ILogger<NewUpdateHandler> logger,
-        AfkService afkService,
-        AntiSpamService antiSpamService,
-        MataService mataService,
-        SettingsService settingsService,
-        TelegramService telegramService,
-        WordFilterService wordFilterService
+        TelegramService telegramService
     )
     {
         _logger = logger;
@@ -38,18 +30,13 @@ public class NewUpdateHandler : IUpdateHandler
     {
         await _telegramService.AddUpdateContext(context);
 
-        if (_telegramService.IsUpdateTooOld()) return;
-
-        var floodCheck = await _telegramService.FloodCheckAsync();
-        if (floodCheck.IsFlood) return;
-
         _logger.LogTrace("NewUpdate: {@V}", _telegramService.Update);
 
         // Pre-Task is should be awaited.
-        var preTaskResult = await RunPreTasks();
+        var preTaskResult = await _telegramService.OnUpdatePreTaskAsync();
 
         // Last, do additional task which bot may do
-        RunPostTasks();
+        _telegramService.OnUpdatePostTaskAsync().InBackground();
 
         if (!preTaskResult)
         {
@@ -60,67 +47,5 @@ public class NewUpdateHandler : IUpdateHandler
         _logger.LogDebug("Continue to next Handler");
 
         await next(context, cancellationToken);
-    }
-
-    private async Task<bool> RunPreTasks()
-    {
-        var op = Operation.Begin("Run PreTask for ChatId: {ChatId}", _telegramService.ChatId);
-
-        var hasRestricted = await _telegramService.CheckChatRestriction();
-
-        if (hasRestricted)
-        {
-            return false;
-        }
-
-        await _telegramService.FireAnalyzer();
-
-        var shouldDelete = await _telegramService.ScanMessageAsync();
-
-        var hasSpam = await _telegramService.AntiSpamCheckAsync();
-
-        if (hasSpam.IsAnyBanned)
-        {
-            return false;
-        }
-
-        var hasUsername = await _telegramService.RunCheckUserUsername();
-
-        if (!hasUsername)
-        {
-            return false;
-        }
-
-        var hasPhotoProfile = await _telegramService.RunCheckUserProfilePhoto();
-
-        if (!hasPhotoProfile)
-        {
-            return false;
-        }
-
-        if (shouldDelete)
-        {
-            return false;
-        }
-
-        op.Complete();
-
-        return true;
-    }
-
-    private void RunPostTasks()
-    {
-        var op = Operation.Begin("Run PostTask");
-
-        var nonAwaitTasks = new List<Task>
-        {
-            _telegramService.EnsureChatSettingsAsync(),
-            _telegramService.AfkCheckAsync(),
-            _telegramService.CheckNameChangesAsync()
-        };
-
-        nonAwaitTasks.InBackgroundAll();
-
-        op.Complete();
     }
 }
